Deleted unused controllers Refactored Views (use less js now) Added sending emails via Gmail Added PDF and HTML generator

+++ Models/Answer.cs
@@ -16,7 +16,7 @@
 
     public class IntegerAnswer : Answer
     {
-        public int? IntegerAnswerValue { get; set; }
+        public long? IntegerAnswerValue { get; set; }
     }
 
     public class CheckboxAnswer : Answer
